Add VDC haltoid(), use in MixedGenerator ctor
I think MixedGenerator has an issue with walls blocking corridors when
only a small connection is available between two areas. There's a
workaround in DungeonGeneratorTest, and SerpentMapGenerator seems
unaffected, so it's probably only a bug that affects direct use of
MixedGenerator without specifying a collection of Coord to link
together.