IColorCenter was incorrectly checking the cache
This may have been the cause of a slight drop in framerate despite the flood of Long values ending. Now the "feed" of 4 ints to the custom hasher for the cache correctly checks for existing colors, and the code should be more clear. HallucinateFilter has been adjusted to use the time instead of a variable framerate, like most of the rest of SquidLib's graphics, and EverythingDemo has some bugs fixed in filtering and some filters rearranged.