Fix String-printing TCF methods
They were only printing the first character, and had been since the change to reuse "mut" as the mutable StringBuilder for all the code that TextCellFactory used that expected a CharSequence rather than a char. When given a String, it should just use that as the CharSequence, so now it does.