Cull, cull, cull!
LibGDX's Frustum class is now used to cull any text cells and Glyphs that are off-screen in SparseLayers (it affects SparseTextMap mainly). This more than doubles framerate in SparseLayers when 'c' is pressed to cheat the field of vision and see everything. It should help when exploring large maps in a game that uses SparseLayers, which reduced framerate significantly before.