Support 253-color quantize in SColor
This also adds an image of the Mona Lisa to test the quantization on a painting that features a person, and QuantizeDemo has some added things it can show. The 253-color quantize has changed a little; now it produces this output https://i.imgur.com/X9whFHe.png if given the same input as in the last commit, https://i.imgur.com/TBex3JD.png . The changes reduce the harshness of color transitions on skin tones, but reduce the clarity of areas with mostly gray.