Serious work on PNG8
Once quantization can adapt to different palettes, I may submit the PNG8 encoder to libGDX as a pull request, if they want it. Right now it can reduce an existing image to 253 fixed colors (including one fully transparent) and save it as PNG-8, but being able to choose the palette is important. All the PNG8 code is currently in DetailedWorldMapWriter, but it may be moved.