Speed boost on nextLong(long)
This applies to RNG, LinnormRNG, and anything that extends AbstractRNG and uses its nextLong(long) implementation. The new version is much faster than the previous one when the bound changes frequently and the JVM can't optimize for one set  of values; it also avoids any division or modulus operations. It's a little faster on more typical usage, but reliably so. It may have some bias, on par with nextInt(int) or nextIntHasty(int), which is to say it has very little bias for game purposes.