Celebrate! With RNG default changes
I've committed code every day for the last 4 years, so it's time to celebrate that achievement by changing all the output of RNG and StatefulRNG to a slightly faster RandomnessSource! Hooray! LinnormRNG is actually a nice amount faster and fits in the same feature-set as LightRNG, and it seems to have comparable or better statistical quality to boot (both very good). A bunch of other RandomnessSource impls had their nextLong(long) bounded random generation methods updated to use the fake-128-bit multiplication technique used by RNG, AbstractRNG, and LinnormRNG. That way is quite a lot faster and doesn't call nextLong() an unknowable number of times. MoonwalkRNG and GWTRNG are unchanged, at least, so if you need compatibility with the LightRNG-using code you can probably switch to MoonwalkRNG and get some features to boot.