SquidLayersTest->LightingTest, cleanup
FlowingCaveGenerator was the last place that required an RNG and didn't permit an IRNG, but it needed virtually no changes to allow any IRNG implementation. It is used in LightingTest now.

(My Git client seems to be giving a strange preview without all changes shown, so there may be some mixup between the version on GitHub and the version I wrote...)