Some changes to PNG8
These reduce the amount of dithering that is applied to images that don't have any inexact colors, hopefully so that it doesn't dither at all (I'm not certain that it does). This also adds a missing overload to PNG8.write() so it can take a FileHandle and still specify whether to dither.