Challenging OrderedMap fix
This fix is mostly one line, changing `last` to `order.peek()` in `fixOrder(int, int)`, but finding it took all day. The big seemed to only affect code that uses `putAt()` or code like it, including `alterAt()`, but would have affected other methods and the data of the OrderedMap if it wasn't stopped, until eventually causing out of bounds exceptions in the ordering. OrderedDataTest is mostly mangled right now trying to pinpoint the issue, but now that this is fixed the mangling isn't really necessary. I'll probably go through and remove the first and last fields from OrderedMap since they weren't public and were used inconsistently internally, which caused this bug.