Adjusting some cone FOV things
This uses a variant on atan2 with a range that's more convenient for FOV's usage; it returns floats from 0 to 1 instead of -PI to PI. This allows the check for whether a cell is within a cone of FOV to be a simpler Math.abs()  call instead of a clunkier technique involving floating point modulus. Ripple FOV still needs checking out, but it seems to use the same logic for angle checks as shadowcasting. The shadowcasting FOV has been tested in SparseDemo (not committed) and in FOVTest, working well in both places.