Woah! Fixed SparseLayers bug, made movement smooth!
This looks like https://i.imgur.com/6kf5gDD.gifv in SparseDemo now with a very low time per animation; there's no pause between cells if the framerate is high enough. This was a goal of mine for a long time but I wasn't sure what was preventing it from working... The bug I fixed was related to animationCount going up and sometimes failing to come back down, which would cause input lockup. I did not expect it to smooth out movement/action-timing so much, but I'm pleased it does.