Improve Coord.hashCode(), Dijkstra experiments
It's unclear if DijkstraMap's performance is improved by much with this commit, or even if it degrades slightly, but it gives the possibility of optimizing it further (tomorrow). GreasedRegion isn't as fast as I want it on some things, and as a Collection of Coord I'm not sure how well it does its job relative to an UnorderedSet.