Get most of the AOEs working again
The last commit changed something unrelated to the DijkstraMap update, but that seemed relevant -- it turns out it made AOEs return empty idealLocations() results. This has been fixed in at least Beam, Burst, and Cloud AOEs, but something's still wrong with ConeAOE, and seems to have been for a while.