Fix various ConeAOE-related issues
This affects DijkstraMap, which had issues where onlyPassable cells were being ignored some of the time and moves could end in allied cells. It affects FOV, where some of the conical FOV with Ripple mode would affect a huge span instead of the requested one. And of course it affects ConeAOE.