Starting to update EverythingDemo
The new version will coexist with the original in a different package for now, and will probably replace the original at some point. It uses SparseLayers instead of SquidLayers, is generally cleaner, and uses some newer features. This is the version from TsarStart in SquidLib-Demos (on my account, tommyettinger, not SquidPony ), with some small updates and changes.