Add SteadyProjectileEffect
This is a variant on PanelEffect.ProjectileEffect that takes a specific amount of time to cross into each next cell, instead of a specific duration for the whole effect. It is meant for ranged weapons that would look odd if their projectiles travel slowly into nearby targets and quickly when going after far-away targets, which is how ProjectileEffect still works.