Fix CrossHash.Hive mistakes
I apparently messed up a copy from some C++ source (which I use to test hashes) into CrossHash.Hive, and the version I was using only passed some of SMHasher anyway. This algorithm, when in C, isn't quite as fast as the earlier C version, but it may be faster than the last Java version because of the mistake. It does pass SMHasher in full, and is still GWT-optimized.