Clean up old stuff; close #201
Occasionally some terms in MonsterGen could potentially be null and would still have `.equals()` called on them, even with a constant String argument. The caller and callee are flipped now so the constant gets the method called on it; standard stuff. It also may affect a lock-and-key class in Condition.java, so that's changed too in the same simple way.