Very simple fix to FAIconsTest
A new SquidPanel was being added to a SquidLayers, with the rest of the SquidLayers using Iosevka Slab for a font and the SquidPanel using Font-Awesome. The problem was that the new SquidPanel was full of '\u0000' chars, which render as blocks, and their color was black, so the presence of the extra SquidPanel blocked the Iosevka-using layers behind it (the SquidPanel's contents were solid black, and the icons were rendered on top of that). With two lines, this bizarre issue is fixed.