Change GWTRNG algo
Lathe32RNG passes 32TB of PractRand, but has worsening issues as it nears the end of testing. I suspect that there is a frequency issue that's mostly apparent when reducing the output to some small int range, as in RNG.between(int, int). Here, I've swapped it for Starfish32RNG, which passes PractRand with no anomalies, as well as TestU01's BigCrush in forward and reverse. Starfish is 2-dimensionally equidistributed where Lathe is only 1-dimensionally, so Starfish can return all ints and all longs (except one long) with equal likelihood, while Lathe can only do that for ints. Starfish isn't quite as fast, but the difference is barely measurable on desktop (it can be somewhat significant on GWT, depending on the browser, but it's still at least 10x faster when generating ints than a 64-bit generator like LinnormRNG). There don't seem to be the same frequent-repeat issues in GWTRNG now.