Add more things to Thesaurus
It can get words for numbers, such as numberWord(3) gives "three", for the adjectives relating to a position of  a number (not sure if this is the right term), such as numberAdjective(12) gives "twelfth", and can get both randomly from 2 to some specified number with synonyms like five`noun` or eleven`adj`.

"Frafodh, The Black Pyre of Drabog, bring forth your terrible frenzy against this wretch!" "By the eight sigils of Elar, in the name of Our Adored Maiden, protect us from this murderer!" "With genius beyond my own, I conjure the Tenth Glyph of Erviangarl! The Countless Brood, brandish thy POWER!"