Add some of Mark Overton's RNGs
Mover32 and Mover64 are both, in Overton's terminology, cmr^cmr generators, and Mover32 has a known and fairly good period. Mover64 has a large period but it isn't known other than that it must be at least 2 to the 38 and is no more than 2 to the 128 minus 1 (and is probably above 2 to the 64). They are both extremely fast; only BasicRandom32 is faster than Mover32RNG, but BasicRandom32 fails tests all over. LapRNG or FlapRNG may also be in the running for fastest very-flawed generator. Both Mover32 and Mover64 pass PractRand testing surprisingly well, both through 32TB.