Some fixes in WorldMapGenerator
SphereMapAlt wasn't calculating polar areas correctly when the map zoomed in, but it does now. Most of the octave frequencies for noise have been adjusted to produce better maps (with less random "holes" in the land), and some of the terrain noise calls have less influence on the finished world map (this helps make the map a little more realistic).