Add LowStorageShuffler
This is a pretty neat piece of code, so thanks, Alan Wolfe ( @Atrix256 ), for making this usage of Feistel Networks so accessible. I'm not using MurmurHash because I think the code on the blog really only needed a simpler hash function that has really strong avalanche qualities, and it didn't need to hash more than one item ever.