Fix mistakes in LowStorageShuffler
I used the `>>` operator instead of `>>>` in a few places where I was porting from C, and this needed to be fixed to correct some results. The round() function is a little more involved now, but only 2 rounds need to be used instead of 4 (this is a net win). There's also more docs.