Fix GreasedRegion.randomScatter()
It was using an instance variable instead of a local one, and treating it as if it could be changed all the time. The tally()-related code seems to only be wrong here, since this is the only place that used a local variable called ct but incorrectly applied it to a non-local.