One-line fix to tally bug in GR
If a GreasedRegion was copied with GreasedRegion.copy() or with new GreasedRegion(GreasedRegion), then the tallied flag would be copied but the counts of how many items were in each section would not be copied. This was a simple copy-paste error on my part. It wasn't an issue with remake(GreasedRegion), which I think is the only place that also does that.