Add IRC link, get GWT building again
There's now a SquidLib channel on Freenode IRC, and the Readme has a link.

A main() method wasn't commented out and it used printf, which blocked all compilation on GWT. :angry:  So now it is commented out, and GWT should build again.

The docs are rebuilt for the first time in a little while.