Change Noise.HastyPointHash
The previous one had some issues with bias on the high bits, and it wasn't faster than PointHash anyway. I don't know how the speed is on this HastyPointHash implementation, but it could be faster than PointHash. It has pretty good visual quality, especially on the low bits, but it hasn't been tested with SMHasher (I'm not sure how it could, since it's adapted for points and not the arbitrary byte arrays SMHasher tests on). This is used in WarpWriter currently for 3D voxel placement, and it works well there.