Experimenting with point hashes, again
They may be able to be much simpler when the input is appropriately scrambled into one value, and that is given to determine() from some random number generator.

This seems to work well with the inverse golden-ratio-like constants from the higher-dimensional Roberts sequences, which are being kept in MummyNoise for now.