Change WorldMapGenerator.SphereMap to avoid 4D noise
I was getting smoother, more coherent maps in other map types, and I think the 4D was to blame because without it, SphereMap produces normal continent-shaped noise. FastNoise is also recommended now throughout WorldMapGenerator because it seems fastest currently (not always, but usually close enough if it isn't) and it seems to usually have pretty good quality.