Tiny change to WMG improves maps made with FastNoise
FastNoise has subtle tetrahedral patterns that can appear in noise when only a low number of octaves are relevant, as in the case where the lowest-frequency octave is weighted the highest in world maps. By increasing the frequency of the ridged noise that is used to break up the primary noise for terrain, we cause sufficient "wear and tear" on what would otherwise be smooth, linear coastlines.