Nice improvement to LightingHandler performance
It seems to get FPS rates between the low-500s to the high-700s now in SparseLightingDemo on this machine, where it had between 300 and 500 before (all of these measurements depend on the area being rendered). This mostly does better when there are not many walls, since they block LOS and so prevent calculations from being needed in the unseen cells.