Improve SwapOrNot. Remove locks package. Docs.
SwapOrNotShuffler and SNShuffledIntSequence are faster now and should have comparable or better quality. The squidpony.squidgrid.mapping.locks package is removed because it didn't really work as hoped, and was a lot of code to never really be used. The commit with the short hash b5236b085d can be used to look up the last version with the locks package, if it matters. Docs are rebuilt, and NOTICE.txt has been updated with attribution.