Improve and add to StringKit number-to-hex methods
You can now convert floats and doubles to the hex representation of their bits, which is handy for packed float colors (it prints them as ABGR). The conversion also doesn't produce garbage now, or at least as much garbage, and there's the option to directly append to a StringBuilder without making an intermediate String.