Change WorldMapGenerator to use two ints instead of one long (GWT)
This only affects the internals for now, but will break classes that extend WorldMapGenerator (change `long state` parameter to `int stateA, int stateB`, change seed to seedA and seedB, change cachedState to cacheA and cacheB, and that should be it). More methods may be added that take two ints for state parameters.