Update libGDX to 1.9.9, fix all the stuff that would have broken
A change to the color used by SpriteBatch was the trickiest to handle, since it went from a float packed color with a temporary Color to updating both every time, even though it mostly reads from the float. Unfortunately, the float is package-private and the setter is now heavyweight, so that old Java standby of creating a package with the same name to access one variable needed to be employed. SparseLayers seems significantly faster than SquidLayers now, and I'm not really sure if that's related to 1.9.9 or stems from the Frustum change here earlier.