Linnorm now fails PractRand (eventually), so Diver is the new default
Diver is only a slight change from Linnorm, using an XLCG instead of an LCG for its state and using some slightly faster operations to scramble that state. It is a win all-around compared to Linnorm, since it is faster, passes 32TB of PractRand instead of failing at 16TB (as Linnorm does in the newest PractRand version, though it passed all tests in earlier versions), and has the same qualities otherwise, like state size and period.

The docs have been rebuilt.