Add method to NumberTools, various utilities to PaletteReducer
NumberTools.reversedIntBitsToFloat() is a nice complement to NumberTools.floatToReversedIntBits(); I need to double-check that this works on GWT as intended just in case I screwed something up. PaletteReducer can get random colors distributed in a way that favors colors that are used often in reductions (so it's less likely to get similar colors). It can also reduce single colors, as RGBA8888 ints, as packed float colors, and as Color objects that are mutated in-place.