Remove some irrelevant assets, such as Iosevka Wide
All of the removed assets can either be reproduced by calling `font.width(w).height(h).initBySize()` to make a font stretch, or were unused and inaccessible from DefaultResources anyway. The font in `DefaultResources.getCrispPrintFamily()` has been improved so the letters line up more smoothly; this was mostly the result of calling `.font().setUseIntegerPositions(false)` on the TextFamily, and this may be used again on other distance field and MSDF fonts.