Fix for (Abstract)RNG.randomPortion(). Don't use LinkedList much.
In earlier SquidLib versions we depended on GWT 2.6.x compatibility, which didn't support ArrayDeque, but the current code has been able to use ArrayDeque for a while. It's usually a better option than LinkedList, but sometimes we use ArrayList instead (like when we don't remove at all). There's also some changes to places where an int division was used in a float context, including some places that were clear bugs.