Fixed a world map bug caused by incorrect expectations of FastNoise
Some places expected a low frequency of about 1/32, others a frequency of exactly 1. FastNoise.instance now gives a FastNoise with a frequency of 1, which helps world maps a lot and doesn't really break anything else (things may change in code that uses SquidLib, but it would have been broken like world maps were if it noticed at all... and you can always instantiate a FastNoise with the parameters you want).