Optimize bloccus image. Start updating TextPanel (removing key input).
My goal with TextPanel is to make it more usable in a game without constantly re-creating the object, and I'm not totally sure how it is going to work. The key input did not ever work for me, did not cooperate with custom keybindings, and couldn't be changed from j and k (which are much more often used for movement). I'll check how this works in Epigon.