Some minor cleanup in DividedMazeGenerator; it has worked for a while...
A fix on April 5, 2018 got DividedMazeGenerator working, possibly for the first time  (since it had never been tested or used). Any version since then on JitPack includes the fix.  The bug was that Arrays.asList() produces a List of some custom type that doesn't allow removal. In general code should prefer Maker.makeList() to avoid bimorphic or megamorphic callsite slowdown (when multiple List implementations could all be used at a callsite, that section of code slows way down; this applies to all interfaces with many implementations, but Maker.makeList returns an ArrayList, which is what most List implementations should be anyway).

A confirmation is in GreasedRegionTest, which uses GreasedRegion to flip true and false, printing walls as '#' and floors as `.` .