OrganicMapGenerator should be cleaner/faster now, and less predictable.
The maps it produced before seemed to very frequently produce large open areas in the center and only formed corridors near the edge. This still happens somewhat, but the edge isn't always accessible now, and corridors can form in the center of the map as well. It's a good alternative to have available, with FlowingCaveGenerator probably a better option when a central open area is not desirable.