Update MiniMover64 to version that passes TestU01 and PractRand
The earlier version passed PractRand, but not generally TestU01's BigCrush or even sometimes SmallCrush (when reversed). The newer version occasionally has failures on a small number of tests in TestU01's BigCrush, but is also capable of having runs with no failures at all, and still passes PractRand to 32TB with no anomalies.