Fixes to ThrustAltRNG quality over the long term.
This may break seed reproducibility if you use NumberTools' randomDouble(), randomSignedDouble(), randomFloat(), or randomFloatCurved() methods, or of course if you use ThrustAltRNG. ThrustAltRNG hasn't been the default for a while, and its code isn't copied to too many places, so this shouldn't have much effect. It would probably be a bad idea to depend on ThrustAltRNG for seeded generation because it can't produce a lot of results, and that may cause a bias in the types of content it can generate.