Cleanup and fixes for obscure features of FastNoise
The cellular noise and gradient perturb option were both broken due to a bad point hash that would easily exceed the valid range for the cases it was used. That bad point hash is gone now. The gradient perturb method names have been brought to the Java standard practice, and now the method names start with lower-case letters.