Remove Splash, consolidate usage of Measurement
DijkstraMap.Measurement has now been promoted to its own non-nested enum, Measurement in squidpony.squidgrid. Code that uses DijkstraMap will probably need to be changed, but this allows me to remove several conflicting Measurement classes with less features spread throughout the code. Splash has been removed; the last straw for me was that it wasn't even doing a randomized floodfill, it was just relying on the order of a collection being random enough after the border was hit by some also-not-random "drunks." That and the fact that it used Chebyshev measurement accounted for why it needed the drunks in the first place; Spill doesn't need them and it should be used as a replacement.