Use a different, simpler lightness calculation for color tables
At least on this laptop, the lightness appears to be much more even in http://squidpony.github.io/SquidLib/ColorTableValue.html , especially around 10% of the way into the list (that had bright blue mixed in with almost-black shades of green and red). The calculation is much simpler, using the exponent 2.19921875 from the recommended Adobe RGB conversion and Hunter LAB's square root in place of CIEL*A*B*'s cube root, but also tweaking the proportions of lightness contribution per RGB channel.