Move the useful parts of ThrustAlt32RNG into ThrustAltRNG, delete 32 version
The 32-bit ThrustAlt32RNG was never that useful as an RNG because it could generate so few numbers before repeating (2 to the 32, which can be exhausted in a few seconds), while also keeping all the issues of ThrustAltRNG. Its determine() methods have been renamed and moved to ThrustAltRNG; ThrustAlt32RNG.determineBounded() is now ThrustAltRNG.determineBoundedShort() and additionally is restricted to generating shorts (there's already a method to generate bounded ints) and so uses especially GWT-friendly math.

Docs have been rebuilt.