I think WorldMapView works now. It's limited, but convenient, and works.
However, coolingModifier in WorldMapGenerator was misleadingly named, so it has changed variable name to heatModifier. This may break code that relied on the public fields of WorldMapGenerator, but it's a quick fix. It does the same thing, so just as increased landMod increases land amount, now increased heatMod increases heat amount.