Change SilkRNG to permit small differences in stateB to change the stream
Before, only very large differences (at least a million) would actually cause a change in the stream unless many numbers were requested. This may slow down SilkRNG a little, but it's still one of the fastest that can be tested in HashVisualizer (which seems like a decent real-world test), and it might be the fastest when generating ints. The sequence of random numbers a given SilkRNG state will produce will change now, but GWT should be the same as desktop/Android targets.