Include an up() step at the end of BinaryHeap.down(); test passes now
This is a rather quick-and-dirty fix, but it does fix BinaryHeap while a more complete solution can be figured out. The issue was that remove() swapped the last node with the removed index, and then filtered down(), but the heap property sometimes requires that the last node be filtered higher in the node array with up() near the end of down(). I don't really understand heaps very well, so there may be much better ways to do this.