I really like GearRNG.
It isn't stateful or skipping, but it is quite fast in some benchmarks (particularly BumbleBench, where it's competitive with the fastest generators, but less-so on JMH benchmarks), and it passes PractRand without anomalies. It's much like a simplified version of SilkRNG that uses 64-bit math, but also includes a comparison that rarely triggers a "gear shift" once in about 10 billion generated numbers. SilkRNG also has this "gear shift," but it does it roughly once every 4 billion generated numbers. Finding the right values for GearRNG took a few weeks of testing, since most problems PractRand could find took about a day of testing to discover.