BREAKING FIX: Direction.getCardinalDirection() was incorrect; is now correct
The bug was that getCardinalDirection() would return UP when it should have returned DOWN, and DOWN when it should have returned UP. This had been present so long in SquidLib that it was expected behavior, and various places worked around the flipped y-axis. If your code uses getCardinalDirection() and flips the y argument (or flips the Direction's deltaY), you should stop doing that as of this commit, since that counteracts a bug that is no longer present and so now would cause a bug itself.