Remember the original goal for WorldMapGenerator, to work as a 3D texture?
Now thanks to @mgsx-dev , we have a 3D demo that loads a now-correct world map texture for a sphere. There's  also a "glitch" world map that has intentional symmetry. The output of SphereMap has been changed so it actually produces a map that can be wrapped on a sphere (Smyth Equal-Surface or Craster Equirectangular).