Add a very nice 128-bit-state RNG, GoatRNG.
GoatRNG isn't the absolute fastest RNG out there (in Java anyway; it's one of the best out there when written in C or C++ and compiled with GCC). Its strongest qualities are more in its, well, quality; it passes PractRand literally to the point my computer could not run the test any more, and does so without any anomalies in the output. It's also nicely robust; it allows any long for both states, and doesn't need warmup to get out of any bad initial state.