Change GreasedRegion: flip() to toggle(), add different flip() that reverses axes
This also fixes some long-standing bugs in and(), or(), xor(), andNot(), and notAnd() when used with two GreasedRegions of different size (specifically, different ySections values). I'll be on the lookout for more cases where ySections was used incorrectly, and writing more tests (the test for cellular automata caught this one, oddly enough).