Benchmark earlygrey/simple-graphs -- it is, truly, simple.
The API is pretty comfortable for simple-graphs, but it isn't especially fast. It runs at reliably half the throughput of DijkstraMap, even when simple-graphs uses Chebyshev as a heuristic and DijkstraMap has none. It's even slower against the Indexed AStar in SquidLib now, with between a third and a quarter the throughput on longer paths (60x60 maps), and a fifth to a sixth on tiny paths.  This is actually pretty good news for simples-graphs because it was only just released, and the performance can probably be improved fairly easily.