Maelstrom all over WMG. Also, adjust frequencies and one vital factor.
The vital factor in map generation here is how much otherRidged, a Noise.Ridged (now wrapped in a Noise.Maelstrom) contributes to heat and moisture domain warping. It was contributing way more than it should, 1.0x, and now it's reduced to 0.375x, which lessens the effect of fake-looking "climate rings" on the map.