Fix a bizarre JDK8 incompatibility in 3.0.1; release 3.0.2
It looks like I compiled the SquidLib 3.0.1 release with Java 14 and didn't specify the Maven `release` to target, so NIO buffer usage was subtly incompatible with Java 8. I've included a fix in FilterBatch that may help with the problem, and now SquidLib should build using Java 14 or newer (which improves the docs) but target Java 8 bytecode. This should work with Java 15, but I haven't tried.