Match up the implementations of Noise and FastNoise for ridged.
This is likely to leave world maps unaffected unless they used FastNoise in a ridged fractal mode as their Noise3D implementation (which would not look like a planet, before or after this change). It may have weird effects if lacunarity and gain are not their defaults (this was already true for billow mode).