Long-overdue fix for a SerpentDeepMapGenerator bug
This one was very rare and hard to track down; it was an infinite-loop bug caused by an internal Set of Coord containing one or more positions that couldn't be used as a link between levels, but the loop condition needed all positions to be used, and so it would try to place an invalid staircase forever. There's some other minor cleanup, too.