Fix Android problem with the IRNG interface, maybe?
There was an exceptionally-strange error trying to build an Android project using the latest SquidLib, which I don't do very often. It was something about IRNG.copy() being considered a Java 8 default method, which isn't supported until Android 24 (and SquidLib should be able to support that, probably). I didn't think it was a default method because it doesn't use the `default` keyword, but let's see if this change helps...