ClassicNoise's hashes may have been doing extra unnecessary work.
Some of what I thought was simple int math was being assigned to a long seed, so I don't know if it was actually being computed as 32-bit or 64-bit; this now definitively uses 64-bit. This changes the hashes for ClassicNoise and probably also JitterNoise. The results seem good for ClassicNoise in a demo I wrote quickly.