Add an extra check to NumberToolsTest; the old atan2 fails it, the current passes.
I've been benchmarking this code as I write it, and this version is about the same speed as the last few commits (faster than before), but passes NumberToolsTest. The extra check is the case where both y and x are infinite; the "imuli" approximation doesn't handle that well and returns NaN, where this does handle it correctly.