SquidMouse: docs and minor cleanup.
We don't need to store gridX and gridY as fields, because they are primitive (no allocation saved by having them as fields), and will be easier for the JVM to work with as locals. I really hope no one is using SquidMouse in a concurrent situation, but if they were, the previous way could be a problem. Using `final` doesn't really matter, though it is something IDEA recommends.

I added some docs to some straightforward method names just because build tools like having docs, I guess.