Update Gradle, publish plugin, anim8-gdx.
The new publish plugin means you'll need to build with at minimum Java 11 now; 8 has been on its way out for a while now, so it's not like anyone couldn't see this coming. The maximum version hasn't changed -- because we still target compatibility with Java 7, we can't use Java 20 or newer, and so the max version you can use to build SquidLib is Java 19.